refactor: Advertisement in separate widget
