+++ dristi/lib/src/features/districts/presentation/pages/districts_list_screen.dart
@@ -1,10 +1,10 @@
 import 'package:dristi/src/core/base/base_consumer_stateful_widget.dart';
 import 'package:dristi/src/core/constants/app_assets.dart';
 import 'package:dristi/src/core/constants/app_values.dart';
+import 'package:dristi/src/core/global_widgets/advertisement_image.dart';
 import 'package:dristi/src/core/routes/app_routes.dart';
 import 'package:dristi/src/core/utils/localization_ext.dart';
 import 'package:dristi/src/features/districts/presentation/riverpod/district_provider.dart';
-import 'package:dristi/src/features/home/advertisements/domain/entity/advertisement_entity.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
@@ -116,49 +116,7 @@ class _DistrictsPageState extends BaseConsumerStatefulWidget<DistrictsScreen> {
           right: AppValues.dimen_16.r,
           bottom: AppValues.dimen_10.r,
         ),
-        child: GestureDetector(
-          onTap: () {
-            navigateToWebView(item: AdvertisementEntity.initial());
-          },
-          child: Stack(
-            alignment: Alignment.center,
-            children: [
-              Container(
-                height: AppValues.dimen_60.r,
-                decoration: BoxDecoration(
-                  borderRadius: BorderRadius.circular(AppValues.dimen_16.r),
-                  color: uiColors.background,
-                  image: const DecorationImage(
-                    image: AssetImage(Assets.advertiseBanner),
-                    fit: BoxFit.cover,
-                  ),
-                ),
-              ),
-              Align(
-                alignment: Alignment.centerRight,
-                child: Padding(
-                  padding: EdgeInsets.only(right: AppValues.dimen_20.w),
-                  child: Container(
-                    padding: EdgeInsets.symmetric(
-                      horizontal: AppValues.dimen_8.r,
-                      vertical: AppValues.dimen_8.r,
-                    ),
-                    decoration: BoxDecoration(
-                      color: uiColors.shadow.withOpacity(0.8),
-                      borderRadius: BorderRadius.circular(
-                        AppValues.dimen_10.r,
-                      ),
-                    ),
-                    child: Text(
-                      context.localization.visitNow,
-                      style: appTextStyles.onImageNovaRegular12,
-                    ),
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ),
+        child: const AdvertisementImage(),
       ),
     );
   }
@@ -232,10 +190,6 @@ class _DistrictsPageState extends BaseConsumerStatefulWidget<DistrictsScreen> {
         ),
       ),
     );
-  }
-
-  void navigateToWebView({required AdvertisementEntity item}) {
-    context.pushNamed(AppRoutes.webView, extra: item);
   }
 
   void navigateToDestinationsPage() {
